Cleaned up and changed transform
Cleaned up code, components now need a pointer to a game object to exist, changed transform from a component sub class to a struct
reason: a component can not exist without a game object, a game object can not exist without a transform. if transform where a component it would not be possible to make a game object.