apply feedback
changed code acording to the feedback.
texture2d is no longer a component but a wrapper class for sdl_texture
added render component that is the only component that renders texture2d objects to the screen
textcomponent now makes ands sends a texture2d object to render component
fps componenet calculates the fps and sends it in string form to textcomponet.